Extract relative path functions
